% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{c360todate}
\alias{c360todate}
\title{Convertit un nombre de jours depuis le 01/01/1960 dans un calendrier à 360 jours en une date réelle}
\usage{
c360todate(x)
}
\description{
Convertit un nombre de jours depuis le 01/01/1960 dans un calendrier à 360 jours en une date réelle
}

