% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{champ}
\alias{champ}
\title{A partir d'une data.table DT et d'une liste de "motifs", crée une liste contenant :
- \code{data} : une data.table contenant la variable `id` de DT et identifiant les observations répondant aux conditions de chaque motif
- \code{tableau} : une `table` qui pour chaque motif identifie les effectifs pour chaque année de naissance
- \code{champ} : vecteur identifiant dans DT les observations ne répondant pas aux conditions des motifs}
\usage{
champ(DT, motifs)
}
\description{
A partir d'une data.table DT et d'une liste de "motifs", crée une liste contenant :
- \code{data} : une data.table contenant la variable `id` de DT et identifiant les observations répondant aux conditions de chaque motif
- \code{tableau} : une `table` qui pour chaque motif identifie les effectifs pour chaque année de naissance
- \code{champ} : vecteur identifiant dans DT les observations ne répondant pas aux conditions des motifs
}

