% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cout.R
\name{coefrevalo}
\alias{coefrevalo}
\title{Renvoit le coefficient de revalorisation
pour une période de liquidation et une date de décès données}
\usage{
coefrevalo(periode, deces)
}
\description{
Renvoit le coefficient de revalorisation
pour une période de liquidation et une date de décès données
}

