% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cout.R
\name{creer_stat}
\alias{creer_stat}
\title{A partir d'une data.table contenant deux colonnes :
- strate : une observation par strate
- variable (nom quelconque) : statistique calculée pour chaque strate
Renvoit une liste contenant :
- strates : un vecteur contenant la stat pour chaque strate
- total : un vecteur contenant la somme de la stat pour toutes les strates}
\usage{
creer_stat(DT)
}
\description{
A partir d'une data.table contenant deux colonnes :
- strate : une observation par strate
- variable (nom quelconque) : statistique calculée pour chaque strate
Renvoit une liste contenant :
- strates : un vecteur contenant la stat pour chaque strate
- total : un vecteur contenant la somme de la stat pour toutes les strates
}

