% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{datetoc360}
\alias{datetoc360}
\title{Convertit une date réelle en nombre de jours depuis le 01/01/1960 dans un calendrier à 360 jours}
\usage{
datetoc360(x)
}
\description{
Convertit une date réelle en nombre de jours depuis le 01/01/1960 dans un calendrier à 360 jours
}

