% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{diffm}
\alias{diffm}
\title{Différence en mois entre deux dates
Le paramètre `millesime` détermine si la distance doit être millésimée.
Par exemple si on se sert de cette fonction pour calculer l'âge en mois le 01/01/2010 d'un individu né le 15/02/1946, le résultat sera 766 mois pour un âge révolu et 767 mois pour un âge millésimé.}
\usage{
diffm(debut, fin, millesime = FALSE)
}
\description{
Différence en mois entre deux dates
Le paramètre `millesime` détermine si la distance doit être millésimée.
Par exemple si on se sert de cette fonction pour calculer l'âge en mois le 01/01/2010 d'un individu né le 15/02/1946, le résultat sera 766 mois pour un âge révolu et 767 mois pour un âge millésimé.
}

