% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import.R
\name{import_b}
\alias{import_b}
\title{Importe la table base_gen465052 fournie par le tuteur.
La variable \code{pens_num} est automatiquement renommée en \code{id} et renvoyée en tant que chaîne de caractères. Hormis \code{pens_num}.
Aucune variable n'est particulièrement attendue et les types sont déduits automatiquement par \code{read.csv}.
Cela signifie que, hormis \code{id}, les chaînes de caractères sont converties en facteurs.}
\usage{
import_b(path, format = "csv")
}
\arguments{
\item{path}{Le chemin vers la table.}

\item{format}{Le format de la table ; seul "csv" est implémenté.}
}
\value{
Une data.table contenant les données des retraités de la FPE. Les variables sont de types primitifs et aucune vérification n'est effectuée.
}
\description{
Importe la table base_gen465052 fournie par le tuteur.
La variable \code{pens_num} est automatiquement renommée en \code{id} et renvoyée en tant que chaîne de caractères. Hormis \code{pens_num}.
Aucune variable n'est particulièrement attendue et les types sont déduits automatiquement par \code{read.csv}.
Cela signifie que, hormis \code{id}, les chaînes de caractères sont converties en facteurs.
}
\examples{
b <- import_b("O:\\\\Annee2/stats/Groupe06/source/base_gen465052.csv")
}

