% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import.R
\name{import_s46}
\alias{import_s46}
\title{Importe la table simul_gen46 fournie par le tuteur.
La variable \code{pens_num} est automatiquement renommée en \code{id} et renvoyée en tant que chaîne de caractères.
Les variables à retenir et leur type sont renseignés dans le corps de la fonction.
On sépare des "variables de mesure" le véritable nom de la variable et le numéro de la période.
La table finale présente alors une observation pour un couple unique (\code{id},\code{periode360})}
\usage{
import_s46(path, format = "csv")
}
\arguments{
\item{path}{Le chemin vers la table.}

\item{format}{Le format de la table ; seul "csv" est implémenté.}
}
\value{
Une data.table contenant les données simulées correctement transposées des retraités de la FPE. Les variables sont de types primitifs et aucune vérification n'est effectuée.
}
\description{
Importe la table simul_gen46 fournie par le tuteur.
La variable \code{pens_num} est automatiquement renommée en \code{id} et renvoyée en tant que chaîne de caractères.
Les variables à retenir et leur type sont renseignés dans le corps de la fonction.
On sépare des "variables de mesure" le véritable nom de la variable et le numéro de la période.
La table finale présente alors une observation pour un couple unique (\code{id},\code{periode360})
}
\examples{
s46 <- import_s46("O:\\\\Annee2/stats/Groupe06/source/simul_gen46_new.csv")
}

