% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{kmplot}
\alias{kmplot}
\title{Courbes de survie Kaplan-Meier.}
\usage{
kmplot(data, by = NULL, censure = TRUE, color = "black")
}
\arguments{
\item{data}{une data.table contenant les données ; doit renseigner la variable \code{jouini.agem} (entier).}

\item{by}{permet de représenter plusieurs courbes sur un même graphique en renseignant sous forme de chaine de caractère la variable d'aggrégation dans \code{data} ; NULL si une seule courbe.}

\item{censure}{booléen qui déterminer si les départs doivent être censurés à 779 mois.}

\item{color}{si une seule courbe, détermine la couleur.}
}
\description{
Renvoit un objet de classe \code{ggplot} permetant de représenter des fonctions de survie.
estimées par les méthode de Kaplan-Meier.
Les courbe sont accompagnées de l'aire représentant un intervalle de confiance à 95%.
}
\examples{
kmbrut <- kmplot(b,censure=FALSE,color="red")
kmsante <- kmplot(b[as.logical(champ)],by="sante")
}

