% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{renamecols}
\alias{renamecols}
\title{Renommer facilement des colonnes dans une data.table à partir d'une liste}
\usage{
renamecols(DT, matching)
}
\description{
Renommer facilement des colonnes dans une data.table à partir d'une liste
}

