% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{replace.na}
\alias{replace.na}
\title{Remplacer les valeurs manquantes dans un facteur par un nouveau label}
\usage{
replace.na(f, label)
}
\description{
Remplacer les valeurs manquantes dans un facteur par un nouveau label
}

