% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{stratifier}
\alias{stratifier}
\title{Capture l'appel à une fonction de modélisation
et remplace l'argument \code{data} par ses parties successives
selon la variable strate.
Renvoit une liste de modèles : un pour chaque strate}
\usage{
stratifier(modelecall)
}
\description{
Capture l'appel à une fonction de modélisation
et remplace l'argument \code{data} par ses parties successives
selon la variable strate.
Renvoit une liste de modèles : un pour chaque strate
}

