% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cout.R
\name{tauxrevalo}
\alias{tauxrevalo}
\title{Renvoit le taux de revalorisation à une date donnée.
Taux de revalorisation réelle du 01/01/2006 au 01/04/2011
puis taux de 17,4% chaque premier janvier des années suivantes}
\usage{
tauxrevalo(date)
}
\description{
Renvoit le taux de revalorisation à une date donnée.
Taux de revalorisation réelle du 01/01/2006 au 01/04/2011
puis taux de 17,4% chaque premier janvier des années suivantes
}

